+++ tutorials/hair_geometry/hair_geometry_device.cpp
@@ -62,7 +62,7 @@ Vec3fa sampleSphere(const float u, const float v)
 void convertTriangleMesh(ISPCTriangleMesh* mesh, RTCScene scene_out)
 {
   RTCGeometry geom = rtcNewGeometry (g_device, RTC_GEOMETRY_TYPE_TRIANGLE, RTC_GEOMETRY_SUBTYPE_DEFAULT);
-  for (size_t t=0; t<mesh->numTimeSteps; t++) {
+  for (unsigned int t=0; t<mesh->numTimeSteps; t++) {
     rtcSetSharedGeometryBuffer(geom,RTC_BUFFER_TYPE_VERTEX,t,RTC_FORMAT_FLOAT3,mesh->positions[t],0,sizeof(Vertex),mesh->numVertices);
   }
   rtcSetSharedGeometryBuffer(geom,RTC_BUFFER_TYPE_INDEX,0,RTC_FORMAT_UINT3,mesh->triangles,0,sizeof(ISPCTriangle),mesh->numTriangles);
@@ -75,7 +75,7 @@ void convertTriangleMesh(ISPCTriangleMesh* mesh, RTCScene scene_out)
 void convertHairSet(ISPCHairSet* hair, RTCScene scene_out)
 {
   RTCGeometry geom = rtcNewGeometry (g_device, hair->type, hair->subtype);
-  for (size_t t=0; t<hair->numTimeSteps; t++) {
+  for (unsigned int t=0; t<hair->numTimeSteps; t++) {
     rtcSetSharedGeometryBuffer(geom,RTC_BUFFER_TYPE_VERTEX,t,RTC_FORMAT_FLOAT4,hair->positions[t],0,sizeof(Vertex),hair->numVertices);
   }
   rtcSetSharedGeometryBuffer(geom,RTC_BUFFER_TYPE_INDEX,0,RTC_FORMAT_UINT,hair->hairs,0,sizeof(ISPCHair),hair->numHairs);
@@ -91,7 +91,7 @@ RTCScene convertScene(ISPCScene* scene_in)
   /* create scene */
   RTCScene scene_out = rtcNewScene(g_device);
 
-  for (size_t i=0; i<scene_in->numGeometries; i++)
+  for (unsigned int i=0; i<scene_in->numGeometries; i++)
   {
     ISPCGeometry* geometry = scene_in->geometries[i];
     if (geometry->type == TRIANGLE_MESH)
@@ -300,7 +300,7 @@ Vec3fa renderPixelStandard(float x, float y, const ISPCCamera& camera, RayStats&
 
   Vec3fa color = Vec3fa(0.0f);
   Vec3fa weight = Vec3fa(1.0f);
-  size_t depth = 0;
+  unsigned int depth = 0;
 
   while (true)
   {
@@ -478,7 +478,7 @@ extern "C" void device_render (int* pixels,
     g_accu = (Vec3fa*) alignedMalloc(width*height*sizeof(Vec3fa));
     g_accu_width = width;
     g_accu_height = height;
-    for (size_t i=0; i<width*height; i++)
+    for (unsigned int i=0; i<width*height; i++)
       g_accu[i] = Vec3fa(0.0f);
   }
 
@@ -491,7 +491,7 @@ extern "C" void device_render (int* pixels,
   g_accu_count++;
   if (camera_changed) {
     g_accu_count=0;
-    for (size_t i=0; i<width*height; i++)
+    for (unsigned int i=0; i<width*height; i++)
       g_accu[i] = Vec3fa(0.0f);
   }
